requirements (4).txt
streamlit==1.30.0
langchain==0.1.0
langchain-openai==0.0.5
python-docx==0.8.11
PyPDF2==3.0.1
pydantic==2.5.2
python-dotenv==1.0.0
streamlit-extras==0.3.5
numpy==1.26.2
pandas==2.1.3